refactor: added dashboard settings component and remove forecasting & currency settings
